add download sync
(cherry picked from commit 9597a6a39e0bb3173e9b05228c14d0322abe73ec)

+++ DTO/WebService/NHIMedicineDTO.cs
@@ -13,7 +13,7 @@ namespace WebServiceDTO
         public string Med_EngName { get; set; }
         public float Med_Amount { get; set; }
         public string Med_Unit { get; set; }
-        public decimal Med_Price { get; set; }
+        public double Med_Price { get; set; }
         public string Med_Manufactor { get; set; }
         public string Med_Form { get; set; }
         public string Med_SingleCompound { get; set; }
