+++ DTO/WebService/NHISpecialMedicineDTO.cs
@@ -15,7 +15,7 @@ namespace WebServiceDTO
         public string SpeMed_SmallCategory { get; set; }
         public string SpeMed_ChineseCategory { get; set; }
         public string SpeMed_Unit { get; set; }
-        public decimal SpeMed_NHIPrice { get; set; }
+        public double SpeMed_NHIPrice { get; set; }
         public string SpeMed_Manufactory { get; set; }
         public string SpeMed_DocumentID { get; set; }
         public DateTime SpeMed_StartDate { get; set; }
